# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Red Light Green Light Platform API",
    "version": "1.0.0",
    "description": "Unified public API served through the gateway. Generated from per-service controller annotations — do not edit by hand (run `npm run generate:openapi`).",
    "contact": {
      "name": "Red Light Green Light Support"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "API Gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthResponse": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "type": "object"
      },
...
      "User": {
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "lastName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "systemPermissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/auth/internal/projects/{projectId}/init-admin": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "userId": {
                    "type": "integer"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Admin initialized successfully"
          }
        },
        "summary": "Initialize project admin (Internal)",
        "tags": ["IAM"]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "security": [],
        "summary": "Authenticate user and return tokens",
        "tags": ["Auth"]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Current user ID"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Get current user context",
        "tags": ["Auth"]
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully"
          },
          "401": {
            "description": "Invalid refresh token"
          }
        },
        "security": [],
        "summary": "Refresh access token using refresh token cookie",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/signup": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "User created successfully"
          },
          "400": {
            "description": "User already exists"
          }
        },
        "security": [],
        "summary": "Create a new user account",
        "tags": ["Auth"]
      }
    },
...
    "/api/v1/auth/users/{userId}/projects": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of project IDs"
          }
        },
        "summary": "List project IDs for a user (IAM)",
        "tags": ["IAM"]
      }
    },
    "/api/v1/auth/users/{userId}/projects/{projectId}/permissions": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of permissions"
          }
        },
        "summary": "Get permissions for a user in a project (IAM)",
        "tags": ["IAM"]
      }
    },
...
  }
}